requirements.txt
time
requests
filehash
virustotal-api
progress
